feat: 143 implement EntriesService with context loader, creator, and updater for entry management
