+++ apps/backend/src/application/services/EntriesService/index.ts
@@ -1,0 +1,4 @@
+export { EntriesService } from './entry.service';
+export { EntryCreator } from './entries.creator';
+export { EntryUpdater } from './entries.updater';
+export { EntriesContextLoader } from './entries.context-loader';
